Moved all mocks to Fakes assembly
