+++ src/Microsoft.HttpRepl.Fakes/Mocks/MockCommand.cs
@@ -1,10 +1,14 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Repl;
 using Microsoft.Repl.Commanding;
 using Microsoft.Repl.Parsing;
 
-namespace Microsoft.Repl.Tests.Mocks
+namespace Microsoft.HttpRepl.Fakes.Mocks
 {
     public class MockCommand : ICommand<object, ICoreParseResult>
     {
