+++ src/Microsoft.HttpRepl.Fakes/Mocks/NullPreferences.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using Microsoft.HttpRepl.Preferences;
 using Microsoft.Repl.ConsoleHandling;
 
-namespace Microsoft.HttpRepl.Tests.Mocks
+namespace Microsoft.HttpRepl.Fakes.Mocks
 {
     public class NullPreferences : IPreferences
     {
