+++ src/Microsoft.HttpRepl.Tests/Commands/ClearCommandTests.cs
@@ -7,7 +7,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
 using Microsoft.HttpRepl.Fakes.Mocks;
-using Microsoft.HttpRepl.Tests.Mocks;
 using Microsoft.Repl;
 using Microsoft.Repl.Commanding;
 using Microsoft.Repl.ConsoleHandling;
