+++ src/Microsoft.HttpRepl.Tests/Commands/GetCommandTests.cs
@@ -4,8 +4,8 @@
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
 using Microsoft.HttpRepl.Fakes.Commands;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.Fakes.SampleApi;
-using Microsoft.HttpRepl.Tests.Mocks;
 using Xunit;
 
 namespace Microsoft.HttpRepl.Tests.Commands
