+++ src/Microsoft.HttpRepl.Tests/Commands/HelpCommandTests.cs
@@ -8,7 +8,6 @@ using Microsoft.HttpRepl.Commands;
 using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.FileSystem;
 using Microsoft.HttpRepl.Preferences;
-using Microsoft.HttpRepl.Tests.Mocks;
 using Microsoft.Repl;
 using Microsoft.Repl.Commanding;
 using Microsoft.Repl.ConsoleHandling;
