+++ src/Microsoft.HttpRepl.Tests/Commands/PostCommandTests.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.HttpRepl.Commands;
 using Microsoft.HttpRepl.Fakes.Commands;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.Fakes.SampleApi;
 using Microsoft.HttpRepl.FileSystem;
-using Microsoft.HttpRepl.Tests.Mocks;
 using Xunit;
 
 namespace Microsoft.HttpRepl.Tests.Commands
