+++ src/Microsoft.HttpRepl.Tests/Commands/PrefCommandTests.cs
@@ -6,10 +6,10 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.FileSystem;
-using Microsoft.HttpRepl.Tests.Mocks;
-using Microsoft.HttpRepl.Tests.Preferences;
 using Microsoft.HttpRepl.Preferences;
+using Microsoft.HttpRepl.Tests.Preferences;
 using Microsoft.HttpRepl.UserProfile;
 using Microsoft.Repl.Parsing;
 using Xunit;
