+++ src/Microsoft.HttpRepl.Tests/Commands/SetBaseCommandTests.cs
@@ -7,8 +7,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.Tests.Commands;
-using Microsoft.HttpRepl.Tests.Mocks;
 using Microsoft.Repl.Parsing;
 using Xunit;
