+++ src/Microsoft.HttpRepl.Tests/Commands/SetHeaderCommandTests.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
-using Microsoft.HttpRepl.Tests.Mocks;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.Repl.Parsing;
 using Xunit;
