+++ src/Microsoft.HttpRepl.Tests/Commands/SetSwaggerCommandTests.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Commands;
-using Microsoft.HttpRepl.Tests.Mocks;
+using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.HttpRepl.Resources;
 using Microsoft.Repl.Parsing;
 using Xunit;
