+++ src/Microsoft.Repl.Tests/ShellTests.cs
@@ -3,7 +3,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.HttpRepl.Fakes.Mocks;
 using Microsoft.Repl.Commanding;
-using Microsoft.Repl.Tests.Mocks;
 using Moq;
 using Xunit;
